// src/components/atoms/Typography/Typography.types.ts
export type TypographyVariant =
  | "heading1"
  | "heading2"
  | "heading3"
  | "heading4"
  | "heading5"
  | "subtitle1"
  | "subtitle2"
  | "subtitle3"
  | "subtitle4"
  | "subtitle5"
  | "subtitle6"
  | "paragraph1"
  | "paragraph2"
  | "paragraph3"
  | "paragraph4"
  | "paragraph5"
  | "label1"
  | "label2"
  | "label3"
  | "label4"
  | "caption1"
  | "caption2";

export type TypographyColor =
  | "primary"
  | "secondary"
  | "hint"
  | "disabled"
  | "inverse"
  | "link"
  | "brand"
  | "info"
  | "success"
  | "warning"
  | "danger";